requirements.txt
Flask==3.0.0
bcrypt==4.1.2
psycopg2-binary